requirements.txt
amqp==2.6.0
asgiref==3.2.7
billiard==3.6.3.0
celery==4.4.5
confusable-homoglyphs==3.2.0
Django==3.0.7
django-filebrowser==3.13.1
django-grappelli==2.14.2
django-registration==3.1
django-tinymce4-lite==1.8.0
future==0.18.2
jsmin==2.2.2
kombu==4.6.10
Pillow==7.1.2
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
vine==1.3.0
